fix: add proper TS support for all Box props
