+++ src/components/Box/Box.md
@@ -1,7 +1,7 @@
 Responsive box-model layout component.
 
 ```jsx harmony
-import Text from './Text';
+import Text from '../Text';
 
 <Box p={9} fontSize={4} width={300} color="white" bg="primary200">
   <Text size="medium" color="white" fontSize={4}>
